requirements.txt
python-gitlab==1.8.0
requests==2.21.0
